+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Version 0.16.0 (TBD)
 
 Implemented:
+* [#151](https://github.com/mailgun/kafka-pixy/pull/151) Added formal support
+  for Kafka versions up to v2.0.0.
 * [#144](https://github.com/mailgun/kafka-pixy/pull/144) Added support for
   specifying and receiving Kafka record headers over both the HTTP and gRPC
   produce and consume interfaces.